LoginDao refactorized

+++ src/pl/workreporter/web/beans/security/login/LoginDao.java
@@ -17,8 +17,8 @@ public interface LoginDao {
     String getPasswordHashByEmail(String email);
     CompleteUserDetails loadUserDetails(String login);
     CompleteUserDetails loadUserDetailsByEmail(String email);
-    List<Integer> getManagedSolutions(String login);
-    List<Integer> getManagedSolutionsByEmail(String email);
-    List<Integer> getManagedTeams(String login);
-    List<Integer> getManagedTeamsByEmail(String email);
+    List<Long> getManagedSolutions(String login);
+    List<Long> getManagedSolutionsByEmail(String email);
+    List<Long> getManagedTeams(String login);
+    List<Long> getManagedTeamsByEmail(String email);
 }
